Simulator: change data structure of results writers
Result writers are kept in a unordered hash table, where the key is
the field name and value in an unique pointer to result writer. In this
way every result writer can write different field.

For backward compatibility, result writer can also be added without a
field name, where 'default' is then used, with a warning message.