Refactor Decomposition
Simplify it a lot by making the actual domain decomposition outside of
class constructor. Implement make_decomposition taking id and number of
processes and perform domain decomposition based on that. Do not save
any additional information to Decompositoin what we don't use at the
moment, we will extend this later on.